# Conflicts:
#	StringTing_Android/app/src/main/java/com/string/leeyun/stringting_android/google_play_item_payment.java

+++ StringTing_Android/app/src/main/java/com/string/leeyun/stringting_android/API/Get_today_introduction.java
@@ -45,7 +45,7 @@ public class Get_today_introduction {
     @SerializedName("opened")
     private boolean opened;
     @SerializedName("score")
-    private int score;
+    private double score;
     private int id;
 
     public void setHeight(int height) {
@@ -70,7 +70,7 @@ public class Get_today_introduction {
     }
 
 
-    public int getScore() {
+    public double getScore() {
         return score;
     }
